clean code moving calendar to new component
