+++ frontend/src/components/Room.tsx
@@ -1,85 +1,20 @@
-import React, {useCallback, useEffect, useRef, useState} from "react";
+import React, { useEffect, useState} from "react";
 import {useParams} from "react-router";
 import {useQuery} from "@tanstack/react-query";
 import {RoomData} from "../models/Room";
 import Api from "../Api";
-import {Activity} from "../models/Activity";
-import {Calendar, DateLocalizer, View, Views} from 'react-big-calendar';
+import {Activity, EventData} from "../models/Activity";
 import 'moment/locale/pl';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
-import {dateFnsLocalizer } from "react-big-calendar";
-import format from "date-fns/format";
-import parse from "date-fns/parse";
-import startOfWeek from "date-fns/startOfWeek";
-import getDay from "date-fns/getDay";
-import {pl} from "date-fns/locale";
-import {Lecturer} from "../models/Lecturer";
+import NewCalendar from "./Calendar";
 
 
-// model rezerwacji do wyświetlenia w kalendarzu
-interface EventData {
-    start: Date;
-    end: Date;
-    course_name: Map<string, string>;
-    classtype_name: Map<string, string>;
-    group_number: number;
-    lecturers: Set<Lecturer>;
-    text: string;
-}
-
-interface EventProps {
-    event: EventData;
-    titleAccessor: keyof EventData;
-    startAccessor: keyof EventData;
-    endAccessor: keyof EventData;
-    allDayAccessor: keyof EventData;
-    tooltipAccessor?: keyof EventData;
-    children?: React.ReactNode;
-}
-
-
-// ustawienie zawartości eventu do wyświetlenia w kalendarzu oraz dostasowanie rozmiaru tekstu
-const Event: React.FC<EventProps> = ({ event, titleAccessor, startAccessor, endAccessor, allDayAccessor, tooltipAccessor, children }) => {
-    const eventRef = useRef<HTMLDivElement>(null);
-
-    useEffect(() => {
-        if (eventRef.current) {
-            const eventElement = eventRef.current;
-            const eventHeight = eventElement.offsetHeight;
-            const eventWidth = eventElement.offsetWidth;
-            // const scaleFactor =  Math.min(eventHeight / 100, eventWidth / 180); // dla szerokich niskich
-            const scaleFactor =  Math.max(Math.min(eventHeight / 120, eventWidth / 210), Math.min(eventHeight / 210, eventWidth / 120)) ;
-            const fontSize = 16 * scaleFactor;
-
-            eventRef.current.style.fontSize = `${fontSize}px`;
-        }
-    }, []);
-
-    return (
-        <div className="rbc-event-content text-center grid content-center " ref={eventRef}>
-            {event.classtype_name && event.group_number &&
-                <div className="mb-1.5 rbc-event-location">
-                    { event.classtype_name["pl"]}, gr.{event.group_number}
-                </div>}
-            {event.course_name &&
-                <div className="rbc-event-description">
-                    {event.course_name["pl"]} -
-                </div>}
-            { Array.from(event.lecturers).map((lecturer) => (
-                  <p>{lecturer.first_name} {lecturer.last_name}</p>
-            ))}
-            {children}
-        </div>
-    );
-};
-
-
-type MyCalendarProps = {
-    events: EventData[];
-};
 
 function Room() {
     const {id} = useParams();
+
+    const[currentDate, setCurrendDate] = useState('');
+    const[calendarDate, setCurrentCalendarDate] = useState(new Date());
 
     const getRoomInfo = () => {
         return Api.get(`room/${id}`).then((res) => res.data);
@@ -90,16 +25,22 @@ function Room() {
         data: room,
         isFetching: isRoomFetching,
         refetch: refetchRoom
-    } = useQuery<RoomData>(["room_info"], getRoomInfo, {
+    } = useQuery<RoomData>(["room_info", currentDate], getRoomInfo, {
         refetchOnWindowFocus: false,
         enabled: true
     });
 
     const getRoomActivities = () => {
-        console.log("DATA:" +  currentDate)
-        // return Api.get(`import/activity/${id}`).then((res) => res.data);
-        // return Api.get(`activity/room/${id}/week?startTime=2023-04-17`).then((res) => res.data);
-        return Api.get(`activity/room/${id}/week?startTime=${currentDate}`).then((res) => res.data);
+        console.log("DAta do kalendarza teraz jest taka : "+ calendarDate )
+
+        if (currentDate == ''){
+            var todayDate = new Date().toISOString().slice(0, 10);
+            console.log("DATA getRoomActivities:" +  todayDate)
+            return Api.get(`activity/room/${id}/week?startTime=${todayDate}`).then((res) => res.data);
+        } else {
+            console.log("DATA getRoomActivities:" +  currentDate)
+            return Api.get(`activity/room/${id}/week?startTime=${currentDate}`).then((res) => res.data);
+        }
     }
 
     // dane do zarządzania informacjami nt. rezerwacji danej sali
@@ -109,27 +50,13 @@ function Room() {
         refetch: refetchActivities
     } = useQuery<Activity[]>(["activities"], getRoomActivities, {
         refetchOnWindowFocus: false,
-        enabled: true
+        enabled: true // false
     });
 
     useEffect(() => {
         refetchRoom();
         refetchActivities();
     }, []);
-
-
-    // ustawienie języka
-    const locales = {
-        pl: pl,
-    };
-
-    const localizer = dateFnsLocalizer({
-        format,
-        parse,
-        startOfWeek,
-        getDay,
-        locales,
-    });
 
 
     // zmiana nazwy typu zajęć na krótszą formę
@@ -192,55 +119,15 @@ function Room() {
         return `${event.classtype_name["pl"]}, gr.${event.group_number}\n${event.course_name["pl"]} - \n${lecturers_txt}`;
     }
 
-
-    // const handleNavigate = (newDateRange: Date[], view: View) => {
-    //     if (view === 'week' && newDateRange[0].getDate() > 7) {
-    //         return myFunction();
-    //     }
-    // };
-
-
-
-    // const [date, setDate] = useState(new Date(2015, 3, 1))
-    // const onNavigate = useCallback((newDate) => {
-    //     setDate(newDate);
-    //     console.log("newdate: " + date);
-    //     getRoomActivities();
-    // }, [setDate])
-    //
-    // const onRangeChange = useCallback((range) => {
-    //     console.log(range)
-    //     // window.alert(buildMessage(range))
-    // }, [])
-
-    // const [currentDate, setCurrentDate] = useState<String>('');
-
-
-
-
-
-
-
-
-
-
-
-    let currentDate = '2023-04-10';
-
-    const handleNavigate = useCallback((newDateRange) => {
-        const date = `${newDateRange.getFullYear()}-${newDateRange.getMonth()+1}-${newDateRange.getDate()-6}`
-        console.log("start: " +  date );
-        currentDate = date;
-        // setCurrentDate(date);
-        // getRoomActivities();
-
-        // tutaj zaktualizowanie listy "activities"
-
-        console.log(activities)
-
-    }, []);
-
-
+    const handleNavigate = ((newDateRange) => {
+        const date = `${newDateRange.getFullYear()}-${newDateRange.getMonth()+1}-${newDateRange.getDate()}`
+        // console.log("start: " +  date );
+        setCurrentCalendarDate(new Date(newDateRange.getFullYear(), newDateRange.getMonth(), newDateRange.getDate()))
+        console.log("UPDATED : " + calendarDate)
+        setCurrendDate(date)
+        console.log("update current date: " + currentDate)
+        refetchActivities()
+    });
 
 
 
@@ -275,42 +162,8 @@ function Room() {
                         <div className="flex flex-col my-8">
                             <hr className="h-px my-8 bg-gray-200 border-0 h-0.5 dark:bg-gray-700"/>
                             <div>
-                                <Calendar
-                                    culture={"pl"}
-                                    localizer={localizer}
-                                    events={roomActivities}
-                                    defaultView={Views.WEEK}
-                                    views={[Views.WEEK]}
-                                    components={{
-                                        event: Event
-                                    }}
-                                    step={15}
-                                    // timeslots={4}
-                                    tooltipAccessor={tooltipAccessor}
-
-                                    // defaultDate={defaultRange.start}
-                                    // date={currentRange.length > 0 ? currentRange[0] : defaultRange.start}
-                                    onNavigate={handleNavigate}
-
-                                    // onNavigate={onNavigate}
-                                    // onRangeChange={handleNavigate}
-                                    dayLayoutAlgorithm="no-overlap"
-                                    startAccessor="start"
-                                    endAccessor="end"
-                                    style={{height: 'full'}}
-                                    min={new Date(0, 0, 0, 6, 0, 0)}
-                                    max={new Date(0, 0, 0, 22, 0, 0)}
-                                    messages={{
-                                        allDay: 'Cały dzień',
-                                        previous: 'Poprzedni',
-                                        next: 'Następny',
-                                        today: 'Dziś',
-                                        day: 'Dzień',
-                                        date: 'Data',
-                                        time: 'Czas',
-                                        event: 'Wydarzenie',
-                                    }}
-                                />
+                                <NewCalendar activities={roomActivities} tooltipAccessor={tooltipAccessor}
+                                             handleNavigate={handleNavigate} calendarDate={calendarDate}/>
                             </div>
                             <div>
                                 <p>tutaj dodanie rezerwacji</p>
@@ -321,11 +174,6 @@ function Room() {
             </div>
         </div>
     );
-}
-
-function myFunction() {
-    console.log('Next week button clicked!');
-    return "";
 }
 
 export default Room;
