+++ frontend/src/models/Activity.ts
@@ -12,3 +12,13 @@ export interface Activity{
     room_id: number;
     lecturers: Set<Lecturer>;
 }
+
+export interface EventData {
+    start: Date;
+    end: Date;
+    course_name: Map<string, string>;
+    classtype_name: Map<string, string>;
+    group_number: number;
+    lecturers: Set<Lecturer>;
+    text: string;
+}
